Make _sharedNullsBuffer use an array of Int32 rather than Int8. The aliasing of this type with pointers of different types is confusing emcc, causing globals with no default value to read the wrong memory. (The generated code should be using HEAP8[addr], but it's using HEAP32[addr>>2], addr is not even 0 mod 4.)

+++ source/core/TypeAndDataManager.cpp
@@ -924,7 +924,7 @@ NamedType::NamedType(TypeManagerRef typeManager, const SubString* name, TypeRef 
 //------------------------------------------------------------
 // AggregateType
 //------------------------------------------------------------
-UInt8 AggregateType::_sharedNullsBuffer[kSharedNullsBufferLength];
+UInt32 AggregateType::_sharedNullsBuffer[kSharedNullsBufferLength];
 //------------------------------------------------------------
 Int32 AggregateType::SubElementCount()
 {
